requirements.txt
kafka-python==2.0.2
msgpack==1.0.6
